refactor(tests): split PropRenderer.test.ts into focused per-feature files
PropRenderer.test.ts was getting long and mixed unrelated concerns.
Split into five focused files under test/unit/scene/props/:

  PropRenderer-ceiling.test.ts   — createCeilingLightFixtures (4 tests)
  PropRenderer-wire-rack.test.ts — createWireRackDisplay (3 tests)
  PropRenderer-dividers.test.ts  — createCategoryDivider (3 tests)
  PropRenderer-floor.test.ts     — createFloorMarkers + entrance mat (4 tests)
  PropRenderer-lifecycle.test.ts — clearProps + dispose (2 tests)

Naming regression tests moved into their respective feature files.
No logic changes; 16 tests all pass.
